+++ engine/src/modules/processing_graph/model/node.cpp
@@ -20,7 +20,7 @@
 #include "node.h"
 #include "generated/lib/model/model.h"
 
-std::optional<std::shared_ptr<NodePort>> Node::getPortById(int32_t id) {
+std::optional<std::shared_ptr<NodePort>> Node::getPortById(int64_t id) {
   for (auto& port : *this->audioInputPorts()) {
     if (port->id() == id) {
       return port;
